requirements.txt
django
django-cors-headers
djangorestframework
django-shopify-auth
Gunicorn
whitenoise
psycopg2